Final Project/preprocess/pyproject.toml
[tool.poetry]
name = "preprocess"
version = "0.1.0"
description = ""
authors = ["Joshua <jschmidtnj@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
